[move-prover] Fixing issues with reference parameters for function values (#17907)
* [move-prover] Fixing issues with reference parameters for function values

This fixes #17420 and additional problems when function values which take references as parameters are used. See the test cases for what is working now.

There is one open problem which has been documented in #17904. This case passes compilation now, just cannot be verified. (It's hopefully a corner case for which we don't need a solution right away.)

This also improves error message on function value use in spec expressions, which is still not supported.

* Addressing reviewer comments

Downstreamed-from: 5e6f3f67f883435295e1171aabc6c1ae8e9659d2