[move value annotator] Cache FatType computation (#18013)
* [move value annotator] Cache FatType computation

This caches the conversion from `StructTag` to `FatStructTag` inside of the annotator. It also wraps an Rc around the `FatStructTag` to avoid cloning when taken from the cache. Computation of `FatStructTag` is supposedly the bottleneck in value annotation, because it needs to fetch transitively all struct definitions from the bytecode. However, it isn't clear yet whether the cache and annotator lifetime is long enough to have a visible effect.

* Addressing reviewer comments

Downstreamed-from: c5fd2db096bc6ede816db3a046dab58ddcc963ee