[compiler] Avoid jumps to pure exit blocks (#18054)
* [compiler] Avoid jumps to pure exit blocks

This is extends the `ControlFlowGraphSimplifier` to optimize jumps to blocks which simply exit the function into the source block. This
simplification may result in dead code which is removed in later stages. Example

```
    L1: ...; goto L2;
    L2: return|abort

    ==>

    L1: ...; return|abort;
    L2: return|abort       // might be dead code now

```

Adding a test demonstrating the effect for matches

* Addressing reviewer comments

Downstreamed-from: c6ae38ec4156111b3c14951f8665cdb3ec87b984